+++ pkg/distributor/distributor.go
@@ -45,6 +45,7 @@ import (
 	ingester_client "github.com/grafana/mimir/pkg/ingester/client"
 	"github.com/grafana/mimir/pkg/mimirpb"
 	"github.com/grafana/mimir/pkg/util"
+	"github.com/grafana/mimir/pkg/util/ephemeral"
 	"github.com/grafana/mimir/pkg/util/globalerror"
 	"github.com/grafana/mimir/pkg/util/httpgrpcutil"
 	util_math "github.com/grafana/mimir/pkg/util/math"
@@ -117,6 +118,8 @@ type Distributor struct {
 	inflightPushRequests      atomic.Int64
 	inflightPushRequestsBytes atomic.Int64
 
+	ephemeralCheckerByUser ephemeral.SeriesCheckerByUser
+
 	// Metrics
 	queryDuration                    *instrument.HistogramCollector
 	ingesterChunksDeduplicated       prometheus.Counter
@@ -177,6 +180,9 @@ type Config struct {
 
 	// Configuration for forwarding of metrics to alternative ingestion endpoint.
 	Forwarding forwarding.Config
+
+	// Enable the experimental feature to mark series as ephemeral.
+	EphemeralSeriesEnabled bool `yaml:"ephemeral_series_enabled" category:"experimental"`
 }
 
 type InstanceLimits struct {
@@ -197,6 +203,7 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet, logger log.Logger) {
 	f.Float64Var(&cfg.InstanceLimits.MaxIngestionRate, maxIngestionRateFlag, 0, "Max ingestion rate (samples/sec) that this distributor will accept. This limit is per-distributor, not per-tenant. Additional push requests will be rejected. Current ingestion rate is computed as exponentially weighted moving average, updated every second. 0 = unlimited.")
 	f.IntVar(&cfg.InstanceLimits.MaxInflightPushRequests, maxInflightPushRequestsFlag, 2000, "Max inflight push requests that this distributor can handle. This limit is per-distributor, not per-tenant. Additional requests will be rejected. 0 = unlimited.")
 	f.IntVar(&cfg.InstanceLimits.MaxInflightPushRequestsBytes, maxInflightPushRequestsBytesFlag, 0, "The sum of the request sizes in bytes of inflight push requests that this distributor can handle. This limit is per-distributor, not per-tenant. Additional requests will be rejected. 0 = unlimited.")
+	f.BoolVar(&cfg.EphemeralSeriesEnabled, "distributor.ephemeral-series-enabled", false, "Enable marking series as ephemeral based on the given matchers in the runtime config.")
 }
 
 // Validate config and returns error on failure
@@ -220,7 +227,7 @@ const (
 )
 
 // New constructs a new Distributor
-func New(cfg Config, clientConfig ingester_client.Config, limits *validation.Overrides, activeGroupsCleanupService *util.ActiveGroupsCleanupService, ingestersRing ring.ReadRing, canJoinDistributorsRing bool, reg prometheus.Registerer, log log.Logger) (*Distributor, error) {
+func New(cfg Config, clientConfig ingester_client.Config, limits *validation.Overrides, activeGroupsCleanupService *util.ActiveGroupsCleanupService, ingestersRing ring.ReadRing, ephemeralChecker ephemeral.SeriesCheckerByUser, canJoinDistributorsRing bool, reg prometheus.Registerer, log log.Logger) (*Distributor, error) {
 	if cfg.IngesterClientFactory == nil {
 		cfg.IngesterClientFactory = func(addr string) (ring_client.PoolClient, error) {
 			return ingester_client.MakeIngesterClient(addr, clientConfig)
@@ -238,14 +245,15 @@ func New(cfg Config, clientConfig ingester_client.Config, limits *validation.Ove
 	subservices = append(subservices, haTracker)
 
 	d := &Distributor{
-		cfg:                   cfg,
-		log:                   log,
-		ingestersRing:         ingestersRing,
-		ingesterPool:          NewPool(cfg.PoolConfig, ingestersRing, cfg.IngesterClientFactory, log),
-		healthyInstancesCount: atomic.NewUint32(0),
-		limits:                limits,
-		HATracker:             haTracker,
-		ingestionRate:         util_math.NewEWMARate(0.2, instanceIngestionRateTickInterval),
+		cfg:                    cfg,
+		log:                    log,
+		ingestersRing:          ingestersRing,
+		ingesterPool:           NewPool(cfg.PoolConfig, ingestersRing, cfg.IngesterClientFactory, log),
+		ephemeralCheckerByUser: ephemeralChecker,
+		healthyInstancesCount:  atomic.NewUint32(0),
+		limits:                 limits,
+		HATracker:              haTracker,
+		ingestionRate:          util_math.NewEWMARate(0.2, instanceIngestionRateTickInterval),
 
 		queryDuration: instrument.NewHistogramCollector(promauto.With(reg).NewHistogramVec(prometheus.HistogramOpts{
 			Namespace: "cortex",
@@ -660,7 +668,7 @@ func (d *Distributor) validateSeries(nowt time.Time, ts mimirpb.PreallocTimeseri
 	}
 
 	if d.limits.MaxGlobalExemplarsPerUser(userID) == 0 {
-		ts.Exemplars = nil
+		mimirpb.ClearExemplars(ts.TimeSeries)
 		return nil
 	}
 
@@ -700,6 +708,7 @@ func (d *Distributor) wrapPushWithMiddlewares(externalMiddleware func(next push.
 	middlewares = append(middlewares, d.prePushRelabelMiddleware)
 	middlewares = append(middlewares, d.prePushValidationMiddleware)
 	middlewares = append(middlewares, d.prePushForwardingMiddleware)
+	middlewares = append(middlewares, d.prePushEphemeralMiddleware)
 	if externalMiddleware != nil {
 		middlewares = append(middlewares, externalMiddleware)
 	}
@@ -1025,6 +1034,61 @@ func (d *Distributor) prePushForwardingMiddleware(next push.Func) push.Func {
 		}
 
 		return resp, httpgrpcutil.PrioritizeRecoverableErr(errs...)
+	}
+}
+
+// prePushEphemeralMiddleware is used as push.Func middleware in front of push method.
+// If marking series as ephemeral is enabled, this middleware uses the ephemeral series
+// provider to determine whether a time series should be marked as ephemeral.
+func (d *Distributor) prePushEphemeralMiddleware(next push.Func) push.Func {
+	if !d.cfg.EphemeralSeriesEnabled {
+		return next
+	}
+
+	return func(ctx context.Context, pushReq *push.Request) (*mimirpb.WriteResponse, error) {
+		userID, err := tenant.TenantID(ctx)
+		if err != nil {
+			return nil, err
+		}
+
+		req, err := pushReq.WriteRequest()
+		if err != nil {
+			return nil, err
+		}
+
+		// Ensure that only time series which shall be marked as ephemeral according to the ephemeral checker get
+		// ingested into the ephemeral store.
+		req.EphemeralTimeseries = nil
+
+		ephemeralChecker := d.ephemeralCheckerByUser.EphemeralChecker(userID, req.Source)
+		if ephemeralChecker != nil {
+			first := true
+			var deleteTs []int
+			for ix := 0; ix < len(req.Timeseries); ix++ {
+				ts := req.Timeseries[ix]
+
+				if !ephemeralChecker.ShouldMarkEphemeral(ts.Labels) {
+					continue
+				}
+
+				if first {
+					req.EphemeralTimeseries = mimirpb.PreallocTimeseriesSliceFromPool()
+					deleteTs = make([]int, 0, len(req.Timeseries)-ix)
+					first = false
+				}
+
+				// Move this series from persistent to ephemeral storage. We don't ingest exemplars for ephemeral series.
+				mimirpb.ClearExemplars(ts.TimeSeries)
+				req.EphemeralTimeseries = append(req.EphemeralTimeseries, ts)
+				deleteTs = append(deleteTs, ix)
+			}
+
+			if len(deleteTs) > 0 {
+				req.Timeseries = util.RemoveSliceIndexes(req.Timeseries, deleteTs)
+			}
+		}
+
+		return next(ctx, pushReq)
 	}
 }